// frontend/types.ts
export interface Image {
  id: number;
  image_url: string;
  alt_text: string;
  caption?: string;
  show_on_home?: boolean;
  image_type: string;
  linked_object_type: string | null;
}
// ...
export interface Certificate {
  id: number;
  title: string;
  issuing_organization: string;
  organization_logo?: string;
  issue_date: string;
  credential_url: string;
  credential_id?: string;
  certificate_image: string;
  description?: string;
  skills: string;
  slug: string;
  images?: Image[];
  show_on_home?: boolean;
}